selection: move to pkg/selection
