+++ pkg/selection/truncation_selection.go
