+++ pkg/selection/truncation_selection_test.go
